% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{ggrf}
\alias{ggrf}
\title{This is a wrap function of randomForest}
\usage{
ggrf(data, ntree = 500, ...)
}
\arguments{
\item{data}{a data frame containg grouping information}

\item{ntree}{the number of trees, default value is 500}

\item{...}{other paras}
}
\value{

}
\description{
This is a wrap function of randomForest
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggrf(data)
}
