% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrf_unsupervise.R
\name{ggrf_uns_plot}
\alias{ggrf_uns_plot}
\title{unsupervised randomForest}
\usage{
ggrf_uns_plot(data, pc = 12, level = 0.68)
}
\arguments{
\item{data}{data frame}

\item{pc}{Which PC to present, one of 12, 13, 23, the default value is 12}

\item{level}{CI default value is 0.68}
}
\value{
a ggplot2 object
}
\description{
unsupervised randomForest
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggrf_uns_plot(data)
}
\author{
Zhonghui Gai
}
