% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrf_unsupervise.R
\name{ggrf_unsupervise}
\alias{ggrf_unsupervise}
\title{unsupervised randomForest}
\usage{
ggrf_unsupervise(data)
}
\arguments{
\item{data}{data frame}
}
\value{
a data frame
}
\description{
unsupervised randomForest
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggrf_unsupervise(data)
}
