% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf_importance}
\alias{rf_importance}
\title{Extract the importance information of randomForest result}
\usage{
rf_importance(rf, type = 2, simply = TRUE, sort = TRUE)
}
\arguments{
\item{rf}{a randomForest result}

\item{type}{1 or 2, default value is 2 for Gini index}

\item{simply}{rownames to col}

\item{sort}{order Gini index}
}
\value{

}
\description{
Extract the importance information of randomForest result
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggrf(data) |>  rf_importance()
}
\author{
Zhonghui Gai
}
