% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf_importance_plot}
\alias{rf_importance_plot}
\title{Plot the result of rf_importance function}
\usage{
rf_importance_plot(data)
}
\arguments{
\item{data}{the result of rf_importance function, a data frame}
}
\value{
a ggplot2 object
}
\description{
Plot the result of rf_importance function
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggrf(data) |>  rf_importance()  |> groutable::dt2()
ggrf(data) |>  rf_importance() |> rf_importance_plot()
}
\author{
Zhonghui Gai
}
